requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.3
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-livereload-server==0.3.2
django-phonenumber-field==5.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
mock==4.0.3
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.14
Pillow==8.0.1
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
tornado==6.1
urllib3==1.25.8
webencodings==0.5.1
djangorestframework==3.12.4
django-compressor==2.4.1
django-imagekit==4.0.2
twilio==6.62.0